% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_meteospain_cache}
\alias{clear_meteospain_cache}
\title{Clear all cached results}
\usage{
clear_meteospain_cache()
}
\description{
Reset the internal cache used to limit the API requests.
}
\details{
Cached results reduces the number of API requests, but sometimes we need
fresh results without restarting the R session. \code{clear_meteospain_cache} function
reset the cache for the actual R session.
}
