% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data_from}
\alias{get_data_from}
\title{Get meteorological stations data}
\usage{
get_data_from(
  service = c("aemet", "smc", "meteoclimatic", "meteogalicia"),
  options
)
}
\arguments{
\item{service}{Character with the service name (in lower case).}

\item{options}{List with the needed service options. See \code{\link{service_options}} to have more info
about the different services and their options.}
}
\value{
An sf (spatial) object with the stations meteorological data.
}
\description{
Connect and retrieve data from AEMET, SMC and other spanish meteorological stations services.
}
\details{
Depending on the service and the temporal resolution selected, the variables present can change, but all
services have at least temperature values.
}
\section{API limits}{

Some APIs have limits in terms of the data that can be retrieved with one call. For example, AEMET
only serves daily data for 31 days in one query. In case a bigger period is wanted, a loop must
be done (see \code{\link[base]{for}}, \code{\link[base]{lapply}} or \code{\link[purrr]{map}}).
}

\examples{
library(meteospain)
library(keyring)

# AEMET (we need a key)
key_set('aemet', user = 'me')
options_for_aemet <-aemet_options(
  'daily',
  start_date = as.Date('2012-01-01'),
  end_date = as.Date('2012-02-01'),
  api_key = key_get('aemet')
)
get_data_from('aemet', options_for_aemet)

}
