% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_stations_info_from}
\alias{get_stations_info_from}
\title{Get metorological stations info}
\usage{
get_stations_info_from(
  service = c("aemet", "smc", "meteoclimatic", "meteogalicia"),
  api_key = NULL
)
}
\arguments{
\item{service}{Character with the service name (in lower case).}

\item{api_key}{API key in case the service needs one. NULL by default.}
}
\value{
An sf (spatial) object with the stations metadata.
}
\description{
Obtain info and metadata for the available stations in the different services
}
\details{
Depending on the service the metadata available can be different. Also, some services only offer
info for active stations (i.e. AEMET), not historical stations, so some mismatch can occur between
the stations returned by this function and the stations returned by \code{\link{get_data_from}} for
historical dates.
}
\examples{
library(meteospain)
library(keyring)

# AEMET (we need a key)
key_set('aemet', user = 'me')
get_station_info_from('aemet', key_get('aemet'))

}
