% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services_options.R
\name{services_options}
\alias{services_options}
\alias{aemet_options}
\alias{smc_options}
\alias{meteoclimatic_options}
\alias{meteogalicia_options}
\title{Options for meteorological services}
\usage{
aemet_options(
  resolution = c("current_day", "daily"),
  start_date = Sys.Date(),
  end_date = NULL,
  stations = NULL,
  api_key
)

smc_options(
  dates,
  stations = NULL,
  resolution = c("instant", "hourly", "daily", "monthly"),
  api_key
)

meteoclimatic_options(resolution = c("current_day"), stations = NULL)

meteogalicia_options(
  resolution = c("instant", "current_day", "daily", "monthly"),
  start_date = Sys.Date(),
  end_date = NULL,
  stations = NULL
)
}
\arguments{
\item{resolution}{Character indicating the temporal resolution for the data. Services allows different
temporal resolutions that can be present or not in each of them (hourly, daily, monthly, instant).}

\item{start_date}{Date class object with the start date from which start collecting data. Ignored if
resolution is one of \code{current_day} or \code{instant}.}

\item{end_date}{Date class object with the end date from which stop collecting data. If NULL (default),
only the date in \code{start_date} is returned. Ignored if resolution is one of \code{current_day} or
\code{instant}.}

\item{stations}{Character vector with the stations codes from which extract data from. If NULL (default)
all available stations are returned.}

\item{api_key}{Character with the API key. NULL by default as not all services require keys. See API Keys
section for more details.}
}
\value{
A list with the service API options to make the query to obtain the data.
}
\description{
Set the options for accessing the different spanish meteorological services

Set the options for accessing the AEMET service

Set the options for accessing the SMC service

Set the options for accessing the Meteoclimatic service

Set the options for accessing the MeteoGalicia service
}
\section{Resolution}{

Temporal resolutions vary from service to service. Check the "Usage" section to see resolutions available
to each service. Possible values are:
\itemize{
\item{\code{current_day} returns the last 24h of measures.}
\item{\code{instant} returns the last measure available.}
\item{\code{hourly} returns any past date/s in hourly format.}
\item{\code{daily} returns any past date/s with daily aggregation.}
\item{\code{monthly} returns any past date/s with monthly aggregation.}
}
}

\section{Keys}{

Some services (i.e. AEMET, SMC...) require an API key to access the data. The requirements and process
to obtain the key varies from service to service.
\itemize{
\item{AEMET: Visit \link{https://opendata.aemet.es/centrodedescargas/inicio} and follow the instructions
at "Obtencion de API Key".}
\item{SMC: Visit \link{https://apidocs.meteocat.gencat.cat/} and follow the instructions there.}
}
It is not advisable to use the keys directly in any script shared or publicly available (github...), neither
store them in plain text files.One option is using the keyring package
(\link{https://github.com/r-lib/keyring}) for managing and accesing keys.
}

\examples{
library(keyring)
library(meteospain)
library(lubridate)

# setting the key (a prompt will appear in console to supply the API key)
keyring::key_set(service = 'aemet', user = 'me')

# create the options
query_options <- aemet_options(
  dates = as_date(as_date('1990-01-01'):as_date('1990-02-01')),
  resolution = 'daily',
  api = keyring::key_get('aemet', 'me')
)

library(keyring)
library(meteospain)
library(lubridate)

# setting the key (a prompt will appear in console to supply the API key)
keyring::key_set(service = 'smc', user = 'me')

# create the options
query_options <- smc_options(
  dates = as_date(as_date('1990-01-01'):as_date('2020-12-31')),
  resolution = 'daily',
  api = keyring::key_get('smc', 'me')
)

library(meteospain)
library(lubridate)

query_options <- meteoclimatic_options()
query_options
# same as before, but more verbose
query_options <- meteoclimatic_options(stations = 'ES', resolution = 'present_day')
query_options

library(meteospain)
library(lubridate)

# create the options
query_options <- meteogalicia_options(
  dates = as_date(as_date('1990-01-01'):as_date('2020-12-31')),
  resolution = 'daily'
)

}
